% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatworkClient.R
\docType{class}
\name{chatworkClient}
\alias{chatworkClient}
\title{make API client object}
\format{An \code{\link{R6Class}} generator object}
\usage{
chatworkClient
}
\description{
make API client object
}
\section{Fields}{

\describe{
\item{\code{access_token}}{your personal access token to authenticate access}
}}
\section{Methods}{

\describe{
\item{\code{post_messages(message, room_id)}}{This methods post your message to specific room.}
\item{\code{get_messages(room_id, force = FALSE)}}{This methods get messages in specific room.}
}
}
\examples{
\dontrun{
client <- rChatwork::chatworkClient$new(access_token)
client$post_messages(message = 'hello chatwork!', room_id = 12345)
client$get_messages(room_id = 12345, force = TRUE)
}
}
\keyword{datasets}

