% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatworkClient.R
\docType{data}
\name{chatwrokClient}
\alias{chatwrokClient}
\title{make API client object}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
chatwrokClient
}
\arguments{
\item{token}{}
}
\value{
API object
}
\description{
make API client object
}
\keyword{datasets}

